+++ nilearn/signal.py
@@ -11,13 +11,12 @@ import warnings
 
 import numpy as np
 import pandas as pd
-from nilearn._utils.glm import _check_run_sample_masks
+from nilearn._utils import fill_doc, stringify_path
+from nilearn._utils.numpy_conversions import as_ndarray, csv_to_array
+from nilearn._utils.param_validation import _check_run_sample_masks
 from scipy import linalg, signal as sp_signal
 from scipy.interpolate import CubicSpline
 from sklearn.utils import as_float_array, gen_even_slices
-
-from ._utils import fill_doc, stringify_path
-from ._utils.numpy_conversions import as_ndarray, csv_to_array
 
 availiable_filters = ["butterworth", "cosine"]
